requirements.txt
kafka-python
pymongo
python-dotenv
certifi
